% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hippo.R
\name{compute_test_statistic}
\alias{compute_test_statistic}
\title{Conduct feature selection by computing test statistics for each gene}
\usage{
compute_test_statistic(df)
}
\arguments{
\item{df}{pre-processed data frame}
}
\value{
data frame with added columns with test results
}
\description{
Conduct feature selection by computing test statistics for each gene
}
