% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ensg_hgnc}
\alias{ensg_hgnc}
\title{A reference data frame that matches ENSG IDs to HGNC symbols}
\format{
A data frame with 46606 rows and 2 columns
\describe{
  \item{ensg}{Ensembl ENSG IDs}
  \item{hgnc}{HGNC symbols}
}
}
\source{
\url{http://www.biomart.org/}
}
\usage{
ensg_hgnc
}
\description{
A reference data frame that matches ENSG IDs to HGNC symbols
}
\keyword{datasets}
