% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hippo.R
\name{get_data_from_sce}
\alias{get_data_from_sce}
\title{Access data from SCE object}
\usage{
get_data_from_sce(sce)
}
\arguments{
\item{sce}{SingleCellExperiment object}
}
\value{
count matrix
}
\description{
Access data from SCE object
}
\examples{
data(toydata)
get_data_from_sce(toydata)
}
