% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hippo.R
\name{get_hippo}
\alias{get_hippo}
\title{Access hippo object from SingleCellExperiment object.}
\usage{
get_hippo(sce)
}
\arguments{
\item{sce}{SingleCellExperiment object}
}
\description{
Access hippo object from SingleCellExperiment object.
}
