% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hippo.R
\name{hippo}
\alias{hippo}
\title{HIPPO's hierarchical clustering}
\usage{
hippo(
  sce,
  K = 20,
  feature_method = c("zero_inflation", "deviance"),
  clustering_method = c("kmeans", "Seurat", "SC3"),
  z_threshold = 1.5,
  deviance_threshold = 150,
  outlier_proportion = 0.001,
  km_num_embeds = 10,
  km_nstart = 50,
  km_iter.max = 50,
  sc3_n_cores = NA,
  verbose = TRUE
)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{K}{maximum number of clusters}

\item{feature_method}{string, either "zero-inflation" or "deviance"}

\item{clustering_method}{string, one of "kmeans", "Seurat", and "SC3"}

\item{z_threshold}{numeric > 0 as a z-value threshold
for selecting the features}

\item{deviance_threshold}{numeric > 0 as a deviance threshold for
selecting the features when method is "deviance}

\item{outlier_proportion}{numeric between 0 and 1, a cut-off
so that when the proportion of important features reach this
number, the clustering terminates}

\item{km_num_embeds}{number of cell embeddings to use in dimension reduction}

\item{km_nstart}{number of tries for k-means for reliability}

\item{km_iter.max}{number of maximum iterations for kmeans}

\item{sc3_n_cores}{number of cores to use if your method is "SC3"}

\item{verbose}{if set to TRUE, shows progress of the algorithm}
}
\value{
a list of clustering result for each level of k=1, 2, ... K.
}
\description{
HIPPO's hierarchical clustering
}
\examples{
data(toydata)
toydata = hippo(toydata,
          feature_method = "zero_inflation",
          clustering_method = "kmeans",
          K = 4,
          outlier_proportion = 0.00001)
}
