% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hippo.R, R/hippo_shuwen.R
\name{hippo_diagnostic_plot}
\alias{hippo_diagnostic_plot}
\title{Create zero-inflation plot compared to reference Poisson line (e^{-x})}
\usage{
hippo_diagnostic_plot(sce, show_outliers = FALSE, zvalue_thresh = 10)

hippo_diagnostic_plot(sce, show_outliers = FALSE, zvalue_thresh = 10)
}
\arguments{
\item{sce}{SingleCellExperiment object with count matrix}

\item{show_outliers}{boolean to indicate whether to circle the outliers
with given zvalue_thresh}

\item{zvalue_thresh}{a numeric vector of the z value threshold for defining
outliers}
}
\value{
a diagnostic plot that shows proportions of zeroes against gene mean
with zero inflation. Black line is the inverse exponential of gene mean.
Outliers are circled in red.

a diagnostic plot that shows proportions of zeroes against gene mean
with zero inflation. Black line is the inverse exponential of gene mean and
outliers, if included, are circled in red.
}
\description{
Create zero-inflation plot compared to reference Poisson line (e^{-x})

Plots zero proportion against gene mean and with inverse exponential fit
by computing test statistics for each gene
}
\examples{
data(toydata)
hippo_diagnostic_plot(toydata, show_outliers=TRUE, zvalue_thresh = 2)
data(toydata)
hippo_diagnostic_plot(toydata, show_outliers=TRUE, zvalue_thresh = 2)
}
