% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hippo.R
\name{hippo_feature_heatmap}
\alias{hippo_feature_heatmap}
\title{HIPPO's feature heatmap}
\usage{
hippo_feature_heatmap(sce, switch_to_hgnc = FALSE, ref = NA,
  top.n = 50, kk = 2)
}
\arguments{
\item{sce}{SingleCellExperiment object with hippo}

\item{switch_to_hgnc}{if the current gene names are ensemble ids, and would like to switch to hgnc}

\item{ref}{a data frame with columns "hgnc" and "ensg" to match each other, only required when switch_to_hgnc is set to TRUE}

\item{top.n}{number of markers to return}
}
\value{
list of differential expression result
}
\description{
HIPPO's feature heatmap
}
\examples{
library(SingleCellExperiment)
X = matrix(rpois(50000, 3), nrow = 1000) # create random count matrix from poisson(10)
X[X\%in\%c(1,2)] = 0
rownames(X) = paste0('gene',1:1000)
colnames(X) = paste0('cell',1:50)
sce = SingleCellExperiment(assays = list(counts = X)) #create SingleCellExperiment object
sce = hippo(sce, K = 3)
sce = diffexp(sce)
}
