% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hippo.R
\name{hippo_umap_plot}
\alias{hippo_umap_plot}
\title{visualize each round of hippo through UMAP}
\usage{
hippo_umap_plot(sce)
}
\arguments{
\item{sce}{SingleCellExperiment object with hippo and UMAP result in it}
}
\value{
umap plot for each round
}
\description{
visualize each round of hippo through UMAP
}
