% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hippo.R
\name{nb_prob_zero}
\alias{nb_prob_zero}
\title{Expected zero proportion under Negative Binomial}
\usage{
nb_prob_zero(lambda, theta)
}
\arguments{
\item{lambda}{gene mean}

\item{theta}{dispersion parameter, 0 if poisson}
}
\value{
Expected zero proportion under Negative Binomial
}
\description{
Expected zero proportion under Negative Binomial
}
