% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hippo.R
\name{one_level_clustering}
\alias{one_level_clustering}
\title{Each round of clustering (sub-function)}
\usage{
one_level_clustering(subX, z_threshold)
}
\arguments{
\item{subX}{data matrix}

\item{z_threshold}{threshold for feature selection}
}
\value{
kmeans result
}
\description{
Each round of clustering (sub-function)
}
