% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hippo.R
\name{pois_prob_zero}
\alias{pois_prob_zero}
\title{Expected zero proportion under Poisson}
\usage{
pois_prob_zero(lambda)
}
\arguments{
\item{lambda}{numeric vector of means of Poisson}
}
\value{
numeric vector of expected proportion of zeros for each lambda
}
\description{
Expected zero proportion under Poisson
}
\examples{
pois_prob_zero(3)
}
