% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hippo.R
\name{preprocess_heterogeneous}
\alias{preprocess_heterogeneous}
\title{Preprocess UMI data without cell label so that each row contains
information about each gene}
\usage{
preprocess_heterogeneous(X)
}
\arguments{
\item{X}{a matrix object with counts data}
}
\value{
data frame with one row for each gene.
}
\description{
Preprocess UMI data without cell label so that each row contains
information about each gene
}
\examples{
data(toydata)
df = preprocess_heterogeneous(get_data_from_sce(toydata))
}
