% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hippo.R, R/hippo_shuwen.R
\name{preprocess_homogeneous}
\alias{preprocess_homogeneous}
\title{Create data frame with gene-level information for each cell type}
\usage{
preprocess_homogeneous(sce, label)

preprocess_homogeneous(sce, label)
}
\arguments{
\item{sce}{SingleCellExperiment object with counts data}

\item{label}{a numeric or character vector of inferred or known label}
}
\value{
data frame with cell-type specific gene information in each row

data frame with one row for each gene.
}
\description{
Create data frame with gene-level information for each cell type

Preprocess UMI data with inferred or known labels
}
\examples{
data(toydata)
labels = SingleCellExperiment::colData(toydata)$phenoid
df = preprocess_homogeneous(toydata, label = labels)
data(toydata)
labels = SingleCellExperiment::colData(toydata)$phenoid
df = preprocess_homogeneous(toydata, label = labels)
}
