% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toydata}
\alias{toydata}
\title{A sample single cell sequencing data subsetted from Zheng2017}
\format{
Single Cell experiment object with 10,000 genes and 100 cells
}
\source{
\url{https://www.nature.com/articles/ncomms14049}
}
\usage{
toydata
}
\description{
A sample single cell sequencing data subsetted from Zheng2017
}
\keyword{datasets}
