% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hippo.R
\name{visualize_hippo}
\alias{visualize_hippo}
\title{visualize each round of hippo}
\usage{
visualize_hippo(hippo_object)
}
\arguments{
\item{hippo_object}{hippo object from the count data}
}
\description{
visualize each round of hippo
}
