% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hippo.R
\name{zinb_prob_zero}
\alias{zinb_prob_zero}
\title{Expected zero proportion under Negative Binomial}
\usage{
zinb_prob_zero(lambda, theta, pi)
}
\arguments{
\item{lambda}{gene mean}

\item{theta}{dispersion parameter, 0 if zero-inflated poisson}

\item{pi}{zero inflation, 0 if negative binomial}
}
\value{
Expected zero proportion under Zero-Inflated Negative Binomial
}
\description{
Expected zero proportion under Negative Binomial
}
