% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hippo.R
\name{compute_test_statistic}
\alias{compute_test_statistic}
\title{Conduct feature selection by computing test statistics for each gene}
\usage{
compute_test_statistic(df)
}
\arguments{
\item{df}{pre-processed data frame}
}
\value{
data frame with added columns with test results
}
\description{
Conduct feature selection by computing test statistics for each gene
}
\examples{
library(SingleCellExperiment)
library(HIPPO)
X = matrix(rpois(1000, 10), nrow = 100) # create random count matrix from poisson(10)
sce = SingleCellExperiment(assays = list(counts = X)) #create SingleCellExperiment object
label = sample(1:5, size=10, replace = TRUE) # create fake cell type label
label = as.factor(label)
df = preprocess_homogeneous(sce, label = label) #get gene information
df = compute_test_statistic(df)
}
