% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hippo.R
\name{diffexp}
\alias{diffexp}
\title{HIPPO's differential expression}
\usage{
diffexp(hippo_object, top.n = 5, switch_to_hgnc = FALSE, ref = NA)
}
\arguments{
\item{hippo_object}{object of hippo}

\item{top.n}{number of markers to return}

\item{switch_to_hgnc}{if the current gene names are ensemble ids, and would like to switch to hgnc}

\item{ref}{a data frame with columns "hgnc" and "ensg" to match each other, only required when switch_to_hgnc is set to TRUE}
}
\value{
list of differential expression result
}
\description{
HIPPO's differential expression
}
