% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hippo.R
\name{dimension_reduction}
\alias{dimension_reduction}
\title{compute t-SNE or umap of each round of HIPPO}
\usage{
dimension_reduction(sce, method = c("umap", "tsne"))
}
\value{
a data frame of dimension reduction result for each k in 1, ..., K
}
\description{
compute t-SNE or umap of each round of HIPPO
}
