% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hippo.R, R/hippo_shuwen.R
\name{get_hippo}
\alias{get_hippo}
\title{Access hippo object from SingleCellExperiment object.}
\usage{
get_hippo(sce)

get_hippo(sce)
}
\arguments{
\item{sce}{SingleCellExperiment object}
}
\value{
hippo object embedded in SingleCellExperiment object

hippo object embedded in SingleCellExperiment object
}
\description{
Access hippo object from SingleCellExperiment object.

Access hippo object from SingleCellExperiment object.
}
\examples{
data(toydata)
set.seed(20200505)
toydata = hippo(toydata,
          feature_method = "zero_inflation",
          clustering_method = "kmeans",
          K = 4,
          outlier_proportion = 0.00001)
hippo_object = get_hippo(toydata)
data(toydata)
set.seed(20200321)
toydata = hippo(toydata,K = 10,z_threshold = 1,outlier_proportion = 0.01)
hippo_object = get_hippo(toydata)
}
