% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hippo.R
\name{get_hippo}
\alias{get_hippo}
\title{Access hippo object from SingleCellExperiment object.}
\usage{
get_hippo(sce)
}
\arguments{
\item{sce}{SingleCellExperiment object}
}
\value{
hippo object embedded in SingleCellExperiment object
}
\description{
Access hippo object from SingleCellExperiment object.
}
\examples{
data(toydata)
set.seed(20200321)
toydata = hippo(toydata,K = 10,
method = "zero_inflation",z_threshold = 1,outlier_proportion = 0.01)
hippo_object = get_hippo(toydata)
}
