% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hippo.R
\name{get_hippo_diffexp}
\alias{get_hippo_diffexp}
\title{Return hippo_diffexp object}
\usage{
get_hippo_diffexp(sce, k = 1)
}
\arguments{
\item{sce}{SingleCellExperiment object with hippo}

\item{k}{integer round of result of interest}
}
\value{
data frame of differential expression test
}
\description{
Return hippo_diffexp object
}
\examples{
data(toydata)
set.seed(20200505)
toydata = hippo(toydata,
          feature_method = "zero_inflation",
          clustering_method = "kmeans",
          K = 4,
          outlier_proportion = 0.00001)
toydata = hippo_diffexp(toydata, method = "gaus")
result1 = get_hippo_diffexp(toydata)
}
