% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hippo.R
\name{get_hippo_diffexp}
\alias{get_hippo_diffexp}
\title{Return hippo_diffexp object}
\usage{
get_hippo_diffexp(sce, k = 1)
}
\arguments{
\item{sce}{SingleCellExperiment object with hippo}

\item{k}{integer round of result of interest}
}
\value{
data frame of differential expression test
}
\description{
Return hippo_diffexp object
}
\examples{
data(toydata)
set.seed(20200321)
toydata = hippo(toydata,K = 10,
method = "zero_inflation",z_threshold = 1,outlier_proportion = 0.01)
toydata = hippo_diffexp(toydata)
result1 = get_hippo_diffexp(toydata)
}
