% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hippo.R
\name{get_hippo_diffexp}
\alias{get_hippo_diffexp}
\title{Return diffexp object}
\usage{
get_hippo_diffexp(sce, k)
}
\arguments{
\item{sce}{SingleCellExperiment object with hippo}

\item{k}{round of result of interest}
}
\value{
data frame of differential expression test
}
\description{
Return diffexp object
}
\examples{
data(toydata)
set.seed(20200321)
toydata = hippo(toydata,K = 10,z_threshold = 1,outlier_proportion = 0.01)
toydata = diffexp(toydata)
get_hippo_diffexp(1)
}
