% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hippo.R
\name{hippo}
\alias{hippo}
\title{HIPPO's hierarchical clustering}
\usage{
hippo(sce, K = 10, z_threshold = 3, outlier_proportion = 0.01)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{K}{number of clusters to ultimately get}

\item{z_threshold}{threshold for selecting the features}
}
\value{
a list of clustering result for each level of k=1, 2, ... K.
}
\description{
HIPPO's hierarchical clustering
}
