% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hippo.R, R/hippo_shuwen.R
\name{hippo}
\alias{hippo}
\title{HIPPO's hierarchical clustering}
\usage{
hippo(
  sce,
  K = 20,
  z_threshold = 2,
  outlier_proportion = 0.001,
  num_embeds = 10,
  nstart = 10,
  verbose = TRUE
)

hippo(
  sce,
  K = 20,
  z_threshold = 2,
  outlier_proportion = 0.001,
  num_embeds = 10,
  nstart = 10,
  verbose = TRUE
)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{K}{maximum number of clusters to ultimately get}

\item{z_threshold}{a numeric > 0 as the z-value threshold
for selecting features}

\item{outlier_proportion}{a numeric between 0 and 1, a cut-off
so that when the proportion of important features reach this
number, the clustering terminates.}

\item{num_embeds}{number of cell embeddings to use in dimension reduction}

\item{nstart}{number of tries for k-means for reliability}

\item{verbose}{if set to TRUE, it shows progress of the algorithm}

\item{feature_method}{string, either "zero-inflation" or "deviance", for feature
selection method}

\item{clustering_method}{string, one of "kmeans", "louvian", and "clustering"}

\item{deviance_threshold}{numeric > 0 as a deviance threshold for
selecting the features when method is "deviance}

\item{km_num_embeds}{number of cell embeddings to use in dimension reduction}

\item{km_nstart}{number of tries for k-means for reliability}

\item{km_iter.max}{number of maximum iterations for kmeans}
}
\value{
a list of clustering result for each level of k=1, 2, ... K.

a list of clustering result of each cell for each level of k=1, 2,
... K.
}
\description{
HIPPO's hierarchical clustering

HIPPO's hierarchical clustering computed using proportion of zeroes
}
\examples{
data(toydata)
toydata = hippo(toydata,
          feature_method = "zero_inflation",
          clustering_method = "kmeans",
          K = 4,
          outlier_proportion = 0.00001)
data(toydata)
toydata = hippo(toydata,K = 10,z_threshold = 1,outlier_proportion = 0.01)
}
