% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hippo.R
\name{hippo_diagnostic_plot}
\alias{hippo_diagnostic_plot}
\title{Conduct feature selection by computing test statistics for each gene}
\usage{
hippo_diagnostic_plot(sce, show_outliers = FALSE, zvalue_thresh = 10)
}
\arguments{
\item{sce}{SingleCellExperiment object with count matrix}

\item{show_outliers}{boolean to indicate whether to circle the outliers
with given zvalue_thresh}

\item{zvalue_thresh}{a numeric v}
}
\value{
a diagnostic plot that shows genes with zero inflation
}
\description{
Conduct feature selection by computing test statistics for each gene
}
\examples{
data(toydata)
hippo_diagnostic_plot(toydata, show_outliers=TRUE, zvalue_thresh = 2)
}
