% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hippo.R
\name{hippo_diagnostic_plot}
\alias{hippo_diagnostic_plot}
\title{Conduct feature selection by computing test statistics for each gene}
\usage{
hippo_diagnostic_plot(sce, show_outliers = FALSE, zvalue_thresh = 10)
}
\arguments{
\item{sce}{SingleCellExperiment object with count matrix}
}
\value{
a diagnostic plot that shows genes with zero inflation
}
\description{
Conduct feature selection by computing test statistics for each gene
}
