% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hippo.R
\name{hippo_diagnostic_plot}
\alias{hippo_diagnostic_plot}
\title{Create zero-inflation plot compared to reference Poisson line (e^{-x})}
\usage{
hippo_diagnostic_plot(sce, show_outliers = FALSE, zvalue_thresh = 10)
}
\arguments{
\item{sce}{SingleCellExperiment object with count matrix}

\item{show_outliers}{boolean to indicate whether to circle the outliers
with given zvalue_thresh}

\item{zvalue_thresh}{a numeric scalar z-value threshold. Genes with z-value
higher than this threshold are marked as outliers when show_outliers is set
to TRUE}
}
\value{
a diagnostic plot that shows proportions of zeroes against gene mean
with zero inflation. Black line is the inverse exponential of gene mean.
Outliers are circled in red.
}
\description{
Create zero-inflation plot compared to reference Poisson line (e^{-x})
}
\examples{
data(toydata)
hippo_diagnostic_plot(toydata, show_outliers=TRUE, zvalue_thresh = 2)
}
