% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hippo.R
\name{hippo_pca_plot}
\alias{hippo_pca_plot}
\title{visualize each round of hippo through t-SNE}
\usage{
hippo_pca_plot(sce, k = NA)
}
\arguments{
\item{sce}{SincleCellExperiment object with hippo and t-SNE result in it}
}
\value{
ggplot for pca in each round
}
\description{
visualize each round of hippo through t-SNE
}
\examples{
data(toydata)
set.seed(20200321)
toydata = hippo(toydata, K = 10,z_threshold = 1)
hippo_pca_plot(toydata, k = 2:3)
}
