% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hippo.R
\name{hippo_tsne_plot}
\alias{hippo_tsne_plot}
\title{visualize each round of hippo through t-SNE}
\usage{
hippo_tsne_plot(sce)
}
\arguments{
\item{sce}{SincleCellExperiment object with hippo and t-SNE result in it}
}
\value{
tsne plot for each round
}
\description{
visualize each round of hippo through t-SNE
}
\examples{
library(SingleCellExperiment)
library(HIPPO)
X = matrix(rpois(1000, 10), nrow = 100) # create random count matrix from poisson(10)
sce = SingleCellExperiment(assays = list(counts = X)) #create SingleCellExperiment object
sce = hippo(sce)
hippo_tsne_plot(sce)
}
