% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hippo.R
\name{hippo_tsne_plot}
\alias{hippo_tsne_plot}
\title{visualize each round of hippo through t-SNE}
\usage{
hippo_tsne_plot(sce, k = NA, title = "")
}
\arguments{
\item{sce}{SincleCellExperiment object with hippo and t-SNE
result in it}
}
\value{
ggplot object for t-SNE in each round
}
\description{
visualize each round of hippo through t-SNE
}
\examples{
data(toydata)
toydata = hippo(toydata,K = 10,z_threshold = 1,outlier_proportion = 0.01)
toydata = dimension_reduction(toydata, method="tsne")
hippo_tsne_plot(toydata)
}
