% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hippo.R
\name{hippo_umap_plot}
\alias{hippo_umap_plot}
\title{visualize each round of hippo through UMAP}
\usage{
hippo_umap_plot(sce)
}
\arguments{
\item{sce}{SingleCellExperiment object with hippo and UMAP result in it}
}
\value{
umap plot for each round
}
\description{
visualize each round of hippo through UMAP
}
\examples{
library(SingleCellExperiment)
library(HIPPO)
X = matrix(rpois(1000, 10), nrow = 100) # create random count matrix from poisson(10)
sce = SingleCellExperiment(assays = list(counts = X)) #create SingleCellExperiment object
sce = hippo(sce)
hippo_umap_plot(sce)
}
