% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hippo.R, R/hippo_shuwen.R
\name{hippo_umap_plot}
\alias{hippo_umap_plot}
\title{visualize each round of hippo through UMAP}
\usage{
hippo_umap_plot(sce, k = NA, pointsize = 0.5, pointalpha = 0.5, plottitle = "")

hippo_umap_plot(sce, k = NA, pointsize = 0.5, pointalpha = 0.5, plottitle = "")
}
\arguments{
\item{sce}{SingleCellExperiment object with hippo and
UMAP result in it}

\item{k}{number of rounds of clustering for which you would like to see result.
Default is 1 to K}

\item{pointsize}{size of the point for the plot (default 0.5)}

\item{pointalpha}{transparency level of points for the plot (default 0.5)}

\item{plottitle}{title of the resulting plot}
}
\value{
ggplot object for umap in each round

ggplot object for umap in each round
}
\description{
visualize each round of hippo through UMAP

visualize each round of hippo through UMAP
}
\examples{
data(toydata)
set.seed(20200505)
toydata = hippo(toydata,
          method = "zero_inflation",
          K = 4,
          outlier_proportion = 0.00001)
toydata = hippo_dimension_reduction(toydata, method="umap")
hippo_umap_plot(toydata)
data(toydata)
set.seed(20200321)
toydata = hippo(toydata,K = 10,z_threshold = 1,outlier_proportion = 0.01)
toydata = hippo_dimension_reduction(toydata, method="umap")
hippo_umap_plot(toydata)
}
