% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hippo.R
\name{hippo_umap_plot}
\alias{hippo_umap_plot}
\title{visualize each round of hippo through UMAP}
\usage{
hippo_umap_plot(sce)
}
\arguments{
\item{sce}{SingleCellExperiment object with hippo and UMAP result in it}
}
\value{
ggplot object for umap in each round
}
\description{
visualize each round of hippo through UMAP
}
\examples{
library(SingleCellExperiment)
library(HIPPO)
X = matrix(rpois(50000, 3), nrow = 1000) # create random count matrix from poisson(10)
X[X\%in\%c(1,2)] = 0
rownames(X) = paste0('gene',1:1000)
colnames(X) = paste0('cell',1:50)
sce = SingleCellExperiment(assays = list(counts = X)) #create SingleCellExperiment object
sce = hippo(sce, K = 3)
sce = dimension_reduction(sce, method="umap")
hippo_umap_plot(sce)
}
