% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hippo.R
\name{hippo_umap_plot}
\alias{hippo_umap_plot}
\title{visualize each round of hippo through UMAP}
\usage{
hippo_umap_plot(sce, k = NA)
}
\arguments{
\item{sce}{SingleCellExperiment object with hippo and
UMAP result in it}
}
\value{
ggplot object for umap in each round
}
\description{
visualize each round of hippo through UMAP
}
\examples{
data(toydata)
toydata = hippo(toydata,K = 10,z_threshold = 1,outlier_proportion = 0.01)
toydata = dimension_reduction(toydata, method="umap")
hippo_umap_plot(toydata)
}
