% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hippo.R
\name{nb_prob_zero}
\alias{nb_prob_zero}
\title{Expected zero proportion under Negative Binomial}
\usage{
nb_prob_zero(lambda, theta)
}
\arguments{
\item{lambda}{numeric vector of means of negative binomial}

\item{theta}{numeric vector of the dispersion parameter
for negative binomial, 0 if poisson}
}
\value{
numeric vector of expected zero proportion under
Negative Binomial
}
\description{
Expected zero proportion under Negative Binomial
}
\examples{
nb_prob_zero(3, 1.1)
}
