% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hippo.R
\name{one_level_clustering}
\alias{one_level_clustering}
\title{Clustering of one-step}
\usage{
one_level_clustering(subX, z_threshold)
}
\arguments{
\item{subX}{a matrix that is gene by cell that needs clustering}

\item{z_threshold}{z-value threshold for feature selection}
}
\value{
a list of clustering result with relevant features, PCA reslut, and kmeans result
}
\description{
Clustering of one-step
}
