% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hippo.R
\name{pois_prob_zero}
\alias{pois_prob_zero}
\title{Expected zero proportion under Poisson}
\usage{
pois_prob_zero(lambda)
}
\arguments{
\item{lambda}{gene mean}
}
\value{
expected proportion of zero under poisson
}
\description{
Expected zero proportion under Poisson
}
