% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hippo_shuwen.R
\name{pois_prob_zero}
\alias{pois_prob_zero}
\title{Expected zero proportion under Poisson}
\usage{
pois_prob_zero(lambda)
}
\arguments{
\item{lambda}{numeric vector of means of Poisson}
}
\value{
Returns a numeric vector of the probability of zeros for the given lambda.
}
\description{
Expected zero proportion under Poisson
}
\examples{
pois_prob_zero(3)
}
