% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hippo.R
\name{preprocess_heterogeneous}
\alias{preprocess_heterogeneous}
\title{Preprocess UMI data without cell label so that each row contains information about each gene}
\usage{
preprocess_heterogeneous(sce)
}
\arguments{
\item{sce}{SingleCellExperiment object with counts data}
}
\value{
data frame with one row for each gene.
}
\description{
Preprocess UMI data without cell label so that each row contains information about each gene
}
\examples{
library(SingleCellExperiment)
library(HIPPO)
X = matrix(rpois(1000, 10), nrow = 100) # create random count matrix from poisson(10)
rownames(X) = paste0('gene',1:100)
colnames(X) = paste0('cell',1:10)
sce = SingleCellExperiment(assays = list(counts = X)) #create SingleCellExperiment object
df = preprocess_heterogeneous(sce) #get gene information
}
