% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hippo.R
\name{preprocess_heterogeneous}
\alias{preprocess_heterogeneous}
\title{Preprocess UMI data so that each row contains information about each gene}
\usage{
preprocess_heterogeneous(sce)
}
\arguments{
\item{X}{gene by cell matrix}
}
\value{
data frame with one row for each gene.
}
\description{
Preprocess UMI data so that each row contains information about each gene
}
