% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hippo.R
\name{preprocess_homogeneous}
\alias{preprocess_homogeneous}
\title{Preprocess UMI data with inferred or known labels}
\usage{
preprocess_homogeneous(sce, label, normalize = FALSE)
}
\arguments{
\item{label}{inferred or known label in factor}

\item{normalize}{normalize each cell to have the same sequencing depth. Default as FALSE}

\item{X}{gene by cell matrix}
}
\value{
data frame with one row for each gene.
}
\description{
Preprocess UMI data with inferred or known labels
}
