% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hippo.R
\name{preprocess_homogeneous}
\alias{preprocess_homogeneous}
\title{Preprocess UMI data with inferred or known labels}
\usage{
preprocess_homogeneous(sce, label, normalize = FALSE)
}
\arguments{
\item{sce}{SingleCellExperiment object with counts data}

\item{label}{a numeric or character vector of inferred or known label}

\item{normalize}{boolean whether to normalize each cell to have the same sequencing depth. Default as FALSE}
}
\value{
data frame with one row for each gene.
}
\description{
Preprocess UMI data with inferred or known labels
}
\examples{
library(SingleCellExperiment)
library(HIPPO)
X = matrix(rpois(2000, 10), nrow = 100) # create random count matrix from poisson(10)
rownames(X) = paste0('gene',1:100)
colnames(X) = paste0('cell',1:20)
sce = SingleCellExperiment(assays = list(counts = X)) #create SingleCellExperiment object
label = sample(1:2, size=20, replace = TRUE) # create fake cell type label
label = as.factor(label)
df = preprocess_homogeneous(sce, label = label) #get gene information
}
