% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hippo.R
\name{preprocess_homogeneous}
\alias{preprocess_homogeneous}
\title{Preprocess UMI data with inferred or known labels}
\usage{
preprocess_homogeneous(sce, label)
}
\arguments{
\item{sce}{SingleCellExperiment object with counts data}

\item{label}{a numeric or character vector of inferred or known label}

\item{normalize}{boolean whether to normalize each cell to have
the same sequencing depth. Default as FALSE}
}
\value{
data frame with one row for each gene.
}
\description{
Preprocess UMI data with inferred or known labels
}
\examples{
data(toydata)
labels = SingleCellExperiment::colData(toydata)$phenoid
df = preprocess_homogeneous(toydata, label = labels)
}
