% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hippo.R
\name{zero_proportion_plot}
\alias{zero_proportion_plot}
\title{visualize each round of hippo through zero proportion plot}
\usage{
zero_proportion_plot(sce)
}
\value{
zero proportion plot
}
\description{
visualize each round of hippo through zero proportion plot
}
