% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hippo.R
\name{zero_proportion_plot}
\alias{zero_proportion_plot}
\title{visualize each round of hippo through zero proportion plot}
\usage{
zero_proportion_plot(sce)
}
\arguments{
\item{sce}{SingleCellExperiment object with hippo element in it}
}
\value{
a ggplot of zero proportion
}
\description{
visualize each round of hippo through zero proportion plot
}
\examples{
library(SingleCellExperiment)
library(HIPPO)
X = matrix(rpois(1000, 10), nrow = 100) # create random count matrix from poisson(10)
sce = SingleCellExperiment(assays = list(counts = X)) #create SingleCellExperiment object
sce = hippo(sce)
zero_proportion_plot(sce)
}
