% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hippo.R
\name{zero_proportion_plot}
\alias{zero_proportion_plot}
\title{visualize each round of hippo through zero proportion plot}
\usage{
zero_proportion_plot(sce, switch_to_hgnc = FALSE, ref = NA, k = NA,
  title = "", topn = 5)
}
\arguments{
\item{sce}{SingleCellExperiment object with hippo element in it}

\item{switch_to_hgnc}{boolean argument to indicate whether to change the
gene names from ENSG IDs to HGNC symbols}

\item{ref}{a data frame with hgnc column and ensg column}
}
\value{
a ggplot object that shows the zero proportions for each round
}
\description{
visualize each round of hippo through zero proportion plot
}
\examples{
data(toydata)
toydata = hippo(toydata,K = 10,z_threshold = 1,outlier_proportion = 0.01)
data(ensg_to_hgnc)
zero_proportion_plot(toydata, switch_to_hgnc = TRUE, ref = ensg_to_hgnc)
}
