% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hippo_shuwen.R
\name{zinb_prob_zero}
\alias{zinb_prob_zero}
\title{Expected zero proportion under Zero-Inflated Negative Binomial}
\usage{
zinb_prob_zero(lambda, theta, pi)
}
\arguments{
\item{lambda}{gene mean, lambda = number of successes r/probability p}

\item{theta}{dispersion parameter, 1/number of
successes r; 0 if zero-inflated poisson}

\item{pi}{zero inflation, 0 if negative binomial}
}
\value{
Expected zero proportion under Zero-Inflated Negative Binomial
}
\description{
Expected zero proportion under Zero-Inflated Negative Binomial
}
\examples{
zinb_prob_zero(3, 1.1, 0.1)
}
